refactor: [헥사고날 아키텍쳐] adapter in, out 패키지에 해당하는 항목 분류
