+++ src/main/java/com/lgmpjt/websnsspringboot/domain/follow/service/FollowPort.java
@@ -1,6 +1,6 @@
 package com.lgmpjt.websnsspringboot.domain.follow.service;
 
-import com.lgmpjt.websnsspringboot.domain.follow.model.Follow;
+import com.lgmpjt.websnsspringboot.adapter.out.entity.Follow;
 
 import java.util.List;
