+++ src/main/java/com/lgmpjt/websnsspringboot/domain/user/service/UserPort.java
@@ -1,6 +1,6 @@
 package com.lgmpjt.websnsspringboot.domain.user.service;
 
-import com.lgmpjt.websnsspringboot.domain.user.model.User;
+import com.lgmpjt.websnsspringboot.adapter.out.entity.User;
 
 public interface UserPort {
 	User save(final User user);
