+++ src/main/java/com/lgmpjt/websnsspringboot/domain/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.lgmpjt.websnsspringboot.domain.user.service;
 
 import com.lgmpjt.websnsspringboot.domain.user.data.UserDto;
-import com.lgmpjt.websnsspringboot.domain.user.model.User;
+import com.lgmpjt.websnsspringboot.adapter.out.entity.User;
 import com.lgmpjt.websnsspringboot.encryption.service.SHA256;
 import com.lgmpjt.websnsspringboot.mapper.user.UserMapper;
 import com.lgmpjt.websnsspringboot.domain.user.data.UserCreateDto;
